Resolve issues with NetApp E-Series LunMapping module (#28294)
* Refactor E-Series LunMapping module to use module_utils

Refactor the NetApp E-Series module to utlize the common module_utils
and doc_fragments.

* Remove the default LUN number

By providing a default LUN number, it interferes with the ability of the
API to determine the appropriate LUN value.

* Fix ignored validate_certs parameter

The validate_certs parameter was not being provided to the underlying
requests method. This patch resolves the issue by passing the value to
all relevant HTTP requests.